Use component file to process & construct component
